refactor: refactored preprocessor and text

+++ SmallTeamSpeech/preprocessor/__init__.py
@@ -5,7 +5,6 @@
     - extracts energy
     - extracts inputs (ex. phonological feats)
 """
-
 from pathlib import Path
 from typing import List, Tuple
 
@@ -13,7 +12,6 @@ import numpy as np
 import pyworld as pw
 import torch  # fix torch imports
 import torchaudio.functional as F
-import torchaudio.transforms as T
 from loguru import logger
 from tabulate import tabulate
 from torch import Tensor, linalg, mean, tensor
@@ -24,7 +22,7 @@ from tqdm import tqdm
 
 from config import ConfigError
 from text import TextProcessor
-from utils import read_textgrid
+from utils import get_spectral_transform, read_textgrid, write_filelist
 
 
 class Preprocessor:
@@ -32,42 +30,31 @@ class Preprocessor:
         self.config = config
         self.missing_files: List[str] = []
         self.audio_config = config["preprocessing"]["audio"]
+        self.sep = config["preprocessing"]["value_separator"]
         self.text_processor = TextProcessor(config)
-        self.feature_prediction_filelist = self.config["training"][
-            "feature_prediction"
-        ]["filelist_loader"](self.config["training"]["feature_prediction"]["filelist"])
-        self.vocoder_filelist = self.config["training"]["vocoder"]["filelist_loader"](
-            self.config["training"]["vocoder"]["filelist"]
-        )
         self.data_dir = Path(self.config["preprocessing"]["data_dir"])
         self.save_dir = Path(self.config["preprocessing"]["save_dir"])
         self.save_dir.mkdir(parents=True, exist_ok=True)
+        self.filelist = self.config["preprocessing"]["filelist_loader"](
+            self.config["preprocessing"]["filelist"]
+        )
         # Define Spectral Transform
         # Gah, so many ways to do this: https://github.com/CookiePPP/VocoderComparisons/issues/3
-        if self.audio_config["spec_type"] == "mel":
-            self.spectral_transform = T.MelSpectrogram(
-                sample_rate=self.audio_config["target_sampling_rate"],
-                n_fft=self.audio_config["n_fft"],
-                win_length=self.audio_config["fft_window_frames"],
-                hop_length=self.audio_config["fft_hop_frames"],
-                n_mels=self.audio_config["n_mels"],
-            )
-        elif config["preprocessing"]["audio"]["spec_type"] == "linear":
-            self.spectral_transform = T.Spectrogram(
-                n_fft=self.audio_config["n_fft"],
-                win_length=self.audio_config["fft_window_frames"],
-                hop_length=self.audio_config["fft_hop_frames"],
-            )
-        elif config["preprocessing"]["audio"]["spec_type"] == "raw":
-            self.spectral_transform = T.Spectrogram(
-                n_fft=self.audio_config["n_fft"],
-                win_length=self.audio_config["fft_window_frames"],
-                hop_length=self.audio_config["fft_hop_frames"],
-                power=None,
-            )
-        else:
+
+        self.spectral_transform = get_spectral_transform(
+            self.audio_config["spec_type"],
+            self.audio_config["n_fft"],
+            self.audio_config["fft_window_frames"],
+            self.audio_config["fft_hop_frames"],
+            sample_rate=self.audio_config["target_sampling_rate"],
+            n_mels=self.audio_config["n_mels"],
+            f_min=self.audio_config["f_min"],
+            f_max=self.audio_config["f_max"],
+        )
+
+        if self.spectral_transform is None:
             raise ConfigError(
-                f"Spectral feature specification {config['preprocessing']['audio']['spec_type'] == 'mel'} is not supported. Please edit your config file."
+                f"Spectral feature specification {self.audio_config['spec_type']} is not supported. Please edit your config file."
             )
 
     def process_audio_for_alignment(self, wav_path) -> Tuple[Tensor, int]:
@@ -106,7 +93,7 @@ class Preprocessor:
         Args:
             audio_tensor (Tensor): Tensor trimmed according
         """
-        return self.spectral_transform(audio_tensor)
+        return self.spectral_transform(audio_tensor).squeeze()
 
     def extract_f0(self, audio_tensor: Tensor):
         """Given an audio tensor, extract the f0
@@ -221,7 +208,7 @@ class Preprocessor:
             spectral_feature_tensor (Tensor): tensor of spectral features extracted from audio
             durations (_type_): _descriptiont    #TODO
         """
-        energy = linalg.norm(spectral_feature_tensor, dim=1)
+        energy = linalg.norm(spectral_feature_tensor, dim=0)
         return energy
 
     def extract_text_inputs(self, text, use_pfs=False):
@@ -255,7 +242,8 @@ class Preprocessor:
 
     def preprocess(
         self,
-        filelist,
+        filelist=None,
+        output_path="processed_filelist.psv",
         process_audio=False,
         process_sox_audio=False,
         process_spec=False,
@@ -264,27 +252,45 @@ class Preprocessor:
         process_duration=False,
         process_pfs=False,
         process_text=False,
+        overwrite=False,
     ):
         # TODO: use multiprocessing
+        write_path = self.save_dir / output_path
+        if write_path.exists and not overwrite:
+            logger.error(
+                f"Preprocessed filelist at '{write_path}' already exists. Please either set overwrite=True or choose a new path"
+            )
+            exit()
         processed = 0
+        files = []
+        if filelist is None:
+            filelist = self.filelist
         for f in tqdm(self.collect_files(filelist), total=len(filelist)):
             speaker = "default" if "speaker" not in f else f["speaker"]
             language = "default" if "language" not in f else f["language"]
+            item = {**f, **{"speaker": speaker, "language": language}}
             audio = None
             spec = None
             if process_text:
                 torch.save(
                     self.extract_text_inputs(f["text"]),
-                    self.save_dir / f"{f['basename']}-{speaker}-{language}-text.npy",
+                    self.save_dir
+                    / self.sep.join([f["basename"], speaker, language, "text.npy"]),
                 )
+                item["raw_text"] = f["text"]
+                item["clean_text"] = self.text_processor.clean_text(f["text"])
             if process_pfs:
                 torch.save(
                     self.extract_text_inputs(f["text"], use_pfs=True),
-                    self.save_dir / f"{f['basename']}-{speaker}-{language}-pfs.npy",
+                    self.save_dir
+                    / self.sep.join([f["basename"], speaker, language, "pfs.npy"]),
                 )
             if process_sox_audio:
                 save_audio(
-                    self.save_dir / f"{f['basename']}-{speaker}-{language}-pfs.npy",
+                    self.save_dir
+                    / self.sep.join(
+                        [f["basename"], speaker, language, "processed.wav"]
+                    ),
                     self.process_audio_for_alignment(
                         self.data_dir / (f["basename"] + ".wav")
                     ),
@@ -298,7 +304,8 @@ class Preprocessor:
                 audio, _ = self.process_audio(self.data_dir / (f["basename"] + ".wav"))
                 torch.save(
                     audio,
-                    self.save_dir / f"{f['basename']}-{speaker}-{language}-audio.npy",
+                    self.save_dir
+                    / self.sep.join([f["basename"], speaker, language, "audio.npy"]),
                 )
             if process_spec:
                 if audio is None:
@@ -309,7 +316,14 @@ class Preprocessor:
                 torch.save(
                     spec,
                     self.save_dir
-                    / f"{f['basename']}-{speaker}-{language}-spec-{self.audio_config['spec_type']}.npy",
+                    / self.sep.join(
+                        [
+                            f["basename"],
+                            speaker,
+                            language,
+                            f"spec-{self.audio_config['spec_type']}.npy",
+                        ]
+                    ),
                 )
             if process_f0:
                 if audio is None:
@@ -318,14 +332,16 @@ class Preprocessor:
                     )
                 torch.save(
                     self.extract_f0(audio),
-                    self.save_dir / f"{f['basename']}-{speaker}-{language}-f0.npy",
+                    self.save_dir
+                    / self.sep.join([f["basename"], speaker, language, "f0.npy"]),
                 )
             if process_energy:
                 if spec is None:
                     spec = self.extract_spectral_features(audio)
                 torch.save(
                     self.extract_energy(spec),
-                    self.save_dir / f"{f['basename']}-{speaker}-{language}-energy.npy",
+                    self.save_dir
+                    / self.sep.join([f["basename"], speaker, language, "energy.npy"]),
                 )
             if process_duration:
                 dur_path = self.data_dir / f["basename"] + ".TextGrid"
@@ -335,7 +351,10 @@ class Preprocessor:
                 torch.save(
                     self.extract_durations(dur_path),
                     self.save_dir
-                    / f"{f['basename']}-{speaker}-{language}-duration.npy",
+                    / self.sep.join([f["basename"], speaker, language, "duration.npy"]),
                 )
+            files.append(item)
             processed += 1
         logger.info(self.report(processed))
+        write_filelist(files, self.save_dir / output_path)
+        return files
